refactor: MovieDetail 매핑 책임 Movie 로 옮김 및 ArrayConverter 추가
